requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-texttospeech
SpeechRecognition
audio-recorder-streamlit